% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnalyzedSigPathway}
\alias{AnalyzedSigPathway}
\title{Information of Analyzed Signature Pathway}
\format{
A data.frame
}
\source{
See \verb{data-raw/}
}
\description{
Information of Analyzed Signature Pathway
}
\examples{
data("AnalyzedSigPathway")
}
