% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apfp}
\alias{apfp}
\title{Information of apfp}
\format{
A data.frame

A data.frame
}
\source{
See \verb{data-raw/}
}
\description{
Information of apfp

Information of brca_disease
}
\examples{
data("apfp")
data("brca_disease")
}
