% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar graphs showed the results of GO and KEGG analysis}
\usage{
bar_plot(
  data,
  top = 10,
  color = "RdBu",
  text.size = 10,
  text.width = 35,
  text.bar = 4,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG and GO results}

\item{top}{according to the order of p adjust value from small to large
the number of categories to show}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{text.size}{text size}

\item{text.width}{y-axis label length}

\item{text.bar}{The size of the text at the top of the bar graph}

\item{title}{title
Kyoto Encyclopedia of Genes and Genomes, KEGG
cellular component, CC
molecular function, MF
biological process, BP}

\item{...}{additional parameters}
}
\value{
barplot
}
\description{
Bar graphs showed the results of GO and KEGG analysis
}
\examples{
\dontrun{
data(xfbdf, package = "TCMR")
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
KK <- setReadable(KK, "org.Hs.eg.db", keyType = "ENTREZID")
BP <- enrichGO(
  gene = eg$ENTREZID,
  "org.Hs.eg.db",
  ont = "BP",
  pvalueCutoff = 0.05,
  readable = TRUE
)
bar_plot(KK,title = "KEGG")
bar_plot(BP,title = "biological process")
}
}
