% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\name{calculate_sig_score}
\alias{calculate_sig_score}
\title{Calculating signature score  on a gene expression dataset}
\usage{
calculate_sig_score(
  eset,
  signature = signature_collection,
  method = "zscore",
  mini_gene_count = 3
)
}
\arguments{
\item{eset}{normalizaed  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures;
such as \code{signature_tme}}

\item{method}{he methods currently supported are \code{ssgsea}, \code{zscore}}

\item{mini_gene_count}{filter out signatures with genes less than minimal gene in expression set;
default is zscore funciion}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating signature score  on a gene expression dataset
}
\examples{
There is no example and please refer to vignette.
}
\references{
\enumerate{
\item Hänzelmann S, Castelo R, Guinney J (2013). “GSVA: gene set variation analysis for microarray and RNA-Seq data.” BMC Bioinformatics, 14, 7. doi: 10.1186/1471-2105-14-7
\item Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates tumour response to PD-L1 blockade by contributing to exclusion of T cells. Nature 554, 544–548 (2018).
}
}
