% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score.R
\name{calculate_sig_score_zscore}
\alias{calculate_sig_score_zscore}
\title{Calculating signature score using z-score method}
\usage{
calculate_sig_score_zscore(eset, signature, mini_gene_count)
}
\arguments{
\item{eset}{transcriptomic data,please make sure a microarray profile or a normalized expression data \link{FPKM or TPM without log2 transformation is recommended} was provided.}

\item{signature}{List of gene signatures}

\item{mini_gene_count}{filter out signatures with genes less than minimal gene in expression set;}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating signature score using z-score method
}
\examples{
There is no example and please refer to vignette.

}
