% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_score.R
\name{calculate_ssgsea}
\alias{calculate_ssgsea}
\title{Calculating drug score using ssGSEA method}
\usage{
calculate_ssgsea(eset, signature)
}
\arguments{
\item{eset}{normalizated  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating drug score using ssGSEA method
}
