% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-score.R
\name{calculate_zscore}
\alias{calculate_zscore}
\title{Calculating drug score using z-score method}
\usage{
calculate_zscore(eset, signature)
}
\arguments{
\item{eset}{transcriptomic data,please make sure a microarray profile or a normalized expression data was provided.}

\item{signature}{List of gene signatures}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating drug score using z-score method
}
