% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cmap}
\alias{cmap}
\title{Title CMAP methods}
\usage{
cmap(input, data, method)
}
\arguments{
\item{input}{input is a list contains up and down genes}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method is cmap}
}
\value{
a data.frame
}
\description{
Title CMAP methods
}
