% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor_tcm}
\alias{cor_tcm}
\title{Title Cor method}
\usage{
cor_tcm(input, data, method)
}
\arguments{
\item{input}{a data.frame contains logFC value, which rownames should be geme symbols}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method from 'pearson',"spearman","kendall"}
}
\value{
a data.frame
}
\description{
Title Cor method
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60,1,drop=FALSE]
aa=cor_tcm(query2,data_logFC,method = 'pearson')
aaa=cor_tcm(query2,data_logFC,method = "spearman")
aaaa=cor_tcm(query2,data_logFC,method = "kendall")
}
