% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreGx_tcm.R
\name{CoreGx_tcm}
\alias{CoreGx_tcm}
\title{Title CoreGx method}
\usage{
CoreGx_tcm(input, data)
}
\arguments{
\item{input}{a data.frame contains logFC value, which rownames should be geme symbols}

\item{data}{a data.frame contains logFC value data from 103 compounds}
}
\value{
a data.frame
}
\description{
Title CoreGx method
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60,1,drop=FALSE]
txp <- CoreGx_tcm(query2,data_logFC[1:3])
}
