% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID19_gene}
\alias{COVID19_gene}
\title{COVID19_gene}
\format{
An object of class \code{character} of length 790.
}
\source{
KEGG & GSEA
}
\usage{
COVID19_gene
}
\description{
A character with 790 genes
}
\examples{
head(COVID19_gene)

}
\keyword{datasets}
