% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cs_plot}
\alias{cs_plot}
\title{Title Visualize Connectivity score}
\usage{
cs_plot(
  data,
  x = "scaled_score",
  y = "tcm",
  colby = "pvalue",
  color = c("blue", "red")
)
}
\arguments{
\item{data}{A data.frame from results of five methods}

\item{x}{x axis}

\item{y}{y axis}

\item{colby}{color sort by a variable}

\item{color}{color to show}
}
\value{
a ggplot
}
\description{
Title Visualize Connectivity score
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
gcmap_kk <- tcm.SSwithGCMAP(input = query2, data = data_logFC)
test <- head(gcmap_kk, 20)
cs_plot(test, x = "raw_score", y = "tcm", colby = "pvalue", color = c("blue", "red"))
}
