% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2t.R
\name{d2t}
\alias{d2t}
\title{How to make disease to target}
\usage{
d2t(x, type = "diseaseID", score = 0.3)
}
\arguments{
\item{x}{is diseaseID or disease}

\item{type}{is the class of x}

\item{score}{is the limitation of score}
}
\value{
target symbol and score
}
\description{
How to make disease to target
}
\examples{
.d2t(x='C0009375',type='diseaseID',score=0.3)
d2t(x=c('C0009375','C0271979'),type='diseaseID',score=0.4)
.d2t(x='Schizophrenia',type='disease',score=0.3)
d2t(x=c('Schizophrenia','beta Thalassemia'),type='disease',score=0.4)
}
