% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_logFC}
\alias{data_logFC}
\title{Information of Analyzed DEG}
\format{
A data.frame
}
\source{
See `data-raw/`
}
\description{
Information of Analyzed DEG
}
\examples{
data("data_logFC")
}
