% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{deg.cal}
\alias{deg.cal}
\title{A Function for Differential Expression Analysis Between two groups}
\usage{
deg.cal(data = expr, group, level, method = "limma")
}
\arguments{
\item{data}{expression dataset}

\item{group}{two groups}

\item{level}{two groups level}

\item{method}{the method for Differential Expression Analysis}
}
\value{
a \code{data.frame}.
}
\description{
A Function for Differential Expression Analysis Between two groups
}
\examples{
data=load_example_dataset()
expr=data$expr[,c(1,2,61,62)]
group_list=group_list_all[c(1,2,61,62)]
tmp=deg.cal(expr,group=group_list,level=group_list[c(3,1)],method='limma')

}
