% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_gene}
\alias{disease_gene}
\title{disease_gene}
\format{
An object of class \code{character} of length 790.
}
\source{
KEGG & GSEA
}
\usage{
disease_gene
}
\description{
Covid 19 gene
A character with 790 genes
}
\examples{
head(disease_gene)

}
\keyword{datasets}
