% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_plot.R
\name{dock_plot}
\alias{dock_plot}
\title{Molecular docking scoring heatmap}
\usage{
dock_plot(
  data,
  shape = NULL,
  color = c("#4393C3", "#F7F7F7", "#D6604D"),
  text.size = 3,
  legend.mid = -2,
  legend.width = 0.5,
  legend.height = 5,
  ...
)
}
\arguments{
\item{data}{Molecular docking results, data frame}

\item{shape}{NULL or "circle"}

\item{color}{color: see "RColorBrewer::display.brewer.all()"}

\item{text.size}{text size}

\item{legend.mid}{legend median}

\item{legend.width}{legend width}

\item{legend.height}{legend height}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Molecular docking scoring heatmap
}
\examples{
data <- matrix(rnorm(81), 9, 9)
data[1:9, seq(1, 9, 2)] <- data[1:9, seq(1, 9, 2)] - 4
colnames(data) <- paste("molecule", 1:9, sep = "")
rownames(data) <- paste("target", 1:9, sep = "")
data <- round(data, 2)
dock_plot(data)
dock_plot(data, shape = "circle", legend.height = 3)
}
