% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_score.R
\name{drug_score}
\alias{drug_score}
\title{Calculating drug score  on a gene expression dataset}
\usage{
drug_score(eset, signature = signature_collection, method = "zscore")
}
\arguments{
\item{eset}{normalizaed  transcriptomic data: normalized (CPM, TPM, RPKM, FPKM, etc.)}

\item{signature}{List of gene signatures;
such as \code{signature_tme}}

\item{method}{the methods currently supported are \code{ssgsea}, \code{zscore}
default is zscore funciion}
}
\value{
data frame with pdata and signature scores for gene sets; signatures in columns, samples in rows
}
\description{
Calculating drug score  on a gene expression dataset
}
\references{
\enumerate{
\item Hänzelmann S, Castelo R, Guinney J (2013). “GSVA: gene set variation analysis for microarray and RNA-Seq data.” BMC Bioinformatics, 14, 7. doi: 10.1186/1471-2105-14-7
\item Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates tumour response to PD-L1 blockade by contributing to exclusion of T cells. Nature 554, 544–548 (2018).
}
}
