% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_similarity.R
\name{drug_similarity}
\alias{drug_similarity}
\title{Title The similarity calculation between two drugs}
\usage{
drug_similarity(smile_a, smile_b, method = "fp", plot = T)
}
\arguments{
\item{smile_a}{the smile of drug a}

\item{smile_b}{the smile of drug b}

\item{method}{the method for calculation}

\item{plot}{a logistical value}
}
\value{
a number
}
\description{
Title The similarity calculation between two drugs
}
\examples{
data("SMILES")
drug_similarity(smile_a = SMILES[1, ], smile_b = SMILES[2, ], method = "fp", plot = TRUE)
}
