% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druglike calculate.R
\name{druglike_calculate}
\alias{druglike_calculate}
\title{Title druglike analysis}
\usage{
druglike_calculate(smiles)
}
\arguments{
\item{smiles}{SMILES of compounds}
}
\value{
a data frame
}
\description{
Title druglike analysis
}
\examples{
data("SMILES")
druglike_calculate(SMILES)
}
