% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{drugpair_seeker}
\alias{drugpair_seeker}
\title{Title Drug Pair Seeker}
\usage{
drugpair_seeker(data = NULL, disease = NULL, cutoff = 1)
}
\arguments{
\item{data}{the logFC matrix of TCM}

\item{disease}{the disease of you intersted}

\item{cutoff}{DEG cutoff}
}
\value{
the data.frame
}
\description{
Title Drug Pair Seeker
}
\examples{
data("data_logFC")
data("tcga_disease")
tmp <- drugpair_seeker(data = data_logFC,disease = res_RNA2)
}
