% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmap_cs.R
\name{gcmap_cs}
\alias{gcmap_cs}
\title{Title gcmap Method}
\usage{
gcmap_cs(input, data, higher = 1, lower = -1)
}
\arguments{
\item{input}{a data.frame contains logFC value, which rownames should be geme symbols}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{higher}{a cutoff value. If logFC larger than or equal to 'higher' will be included in the gene set with  +1}

\item{lower}{a cutoff value. If logFC smaller than or equal to 'higher' will be included in the gene set with  -1, the othres should be set 0.}
}
\value{
a data.frame
}
\description{
Title gcmap Method
}
\examples{
query2 <- data[1:60, 1, drop = FALSE]
gcmap_kk <- gcmap_cs(input = query2, data = data_logFC)
}
