% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dot.R
\name{gene_dot}
\alias{gene_dot}
\title{Title Dotplot of special genes in special compounds of logFC matrix}
\usage{
gene_dot(data, color = c("red", "blue"))
}
\arguments{
\item{data}{the special genes expression data of special compounds}

\item{color}{the color to distinguish between different trends}
}
\value{
A `ggplot` object
}
\description{
Title Dotplot of special genes in special compounds of logFC matrix
}
\examples{
data=as.matrix(data_logFC[1:6,])
data <- data.frame(t(data),check.rows = F,check.names = F)
data$compound <- rownames(data)
data <- data[3:13,]
gene_dot(data)
}
