% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{gene_plot}
\alias{gene_plot}
\title{Title Gene Expression Plot}
\usage{
gene_plot(data, num = 3, colors = c("blue", "white", "red"), shownames = TRUE)
}
\arguments{
\item{data}{Expression data of compounds}

\item{num}{Numbers of cluster}

\item{colors}{Color to show expression level}

\item{shownames}{Logic value to determine whether to display the gene name}
}
\value{
A plot
}
\description{
Title Gene Expression Plot
}
\examples{
gene_plot(data = as.matrix(data_logFC[1:6, ]))
gene_plot(data = as.matrix(data_logFC[1:6, ]), colors = c("blue", "white", "red"))
gene_plot(data = as.matrix(data_logFC[1:10, ]), colors = c("skyblue", "white", "red"))
}
