% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_GO.R
\name{get_sig_GO}
\alias{get_sig_GO}
\title{Get Enriched GO Sets}
\usage{
get_sig_GO(
  data,
  p.cutoff = 0.05,
  p.adj.cutoff = 0.2,
  sortby = "pvalue",
  ont = "BP",
  n.path = 10
)
}
\arguments{
\item{data}{A \code{data.frame} from a subset of \link{deg_caller} result.}

\item{p.cutoff}{the p value cutoff.}

\item{p.adj.cutoff}{the cutoff value of p.adjust.}

\item{sortby}{A string value to indicate the pathway was sorted by p value or p.adjust value. Allowed value contains c('pvalue', 'p.adjust').}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{n.path}{A integer value to indicate how many top pathways sorted by p value should be identified; 10 by default.}
}
\value{
A \code{data.frame}.
}
\description{
Get Enriched GO Sets
}
\examples{
data("AnalyzedDEG")
one_report <- load_analyzedDEG(1)
deg <- subset(one_report, abs(logFC) > 1 & P.Value < 0.05)

go <- get_sig_GO(deg, p.cutoff = 0.05, p.adj.cutoff = 0.5, n.path = 10)
go
dot <- dotplot(go)
dot
bar <- barplot(go)
bar
}
