% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_pathway.R
\name{get_sig_pathway}
\alias{get_sig_pathway}
\title{Get Enriched Signal Pathways}
\usage{
get_sig_pathway(
  data,
  p.cutoff = 0.05,
  p.adj.cutoff = 0.2,
  sortby = "pvalue",
  n.path = 10
)
}
\arguments{
\item{data}{A `data.frame` from a subset of [deg_caller] result.}

\item{p.cutoff}{the p value cutoff.}

\item{p.adj.cutoff}{the cutoff value of p.adjust.}

\item{sortby}{A string value to indicate the pathway was sorted by p value or p.adjust value. Allowed value contains c('pvalue', 'p.adjust').}

\item{n.path}{A integer value to indicate how many top pathways sorted by p value should be identified; 10 by default.}
}
\value{
A `data.frame`.
}
\description{
Get Enriched Signal Pathways
}
\examples{
data("AnalyzedDEG")
one_report <- load_analyzedDEG(1)
deg <- subset(one_report, abs(logFC) > 1 & P.Value < 0.05)
path1 <- get_sig_pathway(deg, p.cutoff = 0.05, p.adj.cutoff = 0.5, n.path = 10)
path2 <- get_sig_pathway(deg, p.cutoff = 0.05, p.adj.cutoff = 0.5, sortby = "p.adjust", n.path = 10)
}
