% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbarplot.R
\name{getbarplot}
\alias{getbarplot}
\title{Barplot of enrichResult}
\usage{
getbarplot(
  data,
  color = c("blue", "red"),
  linetype = "dashed",
  linecol = "chocolate",
  pointcol = "black",
  filcol = "red",
  colby = "pvalue"
)
}
\arguments{
\item{data}{enrichResult object which is \code{data.frame} format.}

\item{color}{color to show}

\item{linetype}{linetype}

\item{linecol}{line color}

\item{pointcol}{point color}

\item{filcol}{bar color}

\item{colby}{mapping colors by pvalue or p.adjust}
}
\value{
ggplot object
}
\description{
Barplot of enrichResult
}
\examples{

data("Analyzedsigpath")
one_report <- load_analyzedsigpath(1)
getbarplot(one_report)
}
