% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdotplot.R
\name{getdotplot}
\alias{getdotplot}
\title{Dotplot for enrichResult}
\usage{
getdotplot(data, fill = "pvalue", color = c("blue", "red"))
}
\arguments{
\item{data}{enrichResult result in `data.frame` format}

\item{fill}{Dot color}

\item{color}{Color to show}
}
\value{
A `ggplot` object
}
\description{
Dotplot for enrichResult
}
\examples{
data("AnalyzedSigPathway")
one_report <- load_AnalyzedSigPathway(1)
getdotplot(one_report)
}
