% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsigpath.R
\name{getsigpath}
\alias{getsigpath}
\title{Get Differential Pathway
\itemize{
\item \code{getsigpath} is used to obtain differential pathway with p < p.cutoff .
}}
\usage{
getsigpath(
  data,
  p.cutoff = 0.05,
  p.adj.cutoff = NULL,
  sortby = "pvalue",
  n.path = 10
)
}
\arguments{
\item{data}{A \code{data.frame} format data from \code{deg_caller} function.}

\item{p.cutoff}{the cutoff value of p.}

\item{p.adj.cutoff}{the cutoff value of p.adjust.}

\item{sortby}{A string value to indicate the pathway was sorted by pvalue or p.adjust value. Allowed value contains c('pvalue', 'p.adjust').}

\item{n.path}{A integer value to indicate how many top pathways sorted by pvalue should be identified; 10 by default.}
}
\value{
A \code{data.frame}.
}
\description{
Get Differential Pathway
\itemize{
\item \code{getsigpath} is used to obtain differential pathway with p < p.cutoff .
}
}
\examples{
data("AnalyzedDEG")
one_report <- load_analyzedDEG(1)
deg <- subset(one_report,abs(logFC)>1&P.Value<0.05)
aa=getsigpath(deg,p.cutoff = 0.05,p.adj.cutoff = 0.5,n.path = 10)
aaa=getsigpath(deg,p.cutoff = 0.05,p.adj.cutoff = 0.5,sortby ='p.adjust' ,n.path = 10)

}
