% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsea}
\alias{gsea}
\title{gsea}
\format{
An object of class \code{data.frame} with 7321 rows and 2 columns.
}
\source{
GSEA database
}
\usage{
gsea
}
\description{
The results of GSEA pathways and genes
A data frame with 7321rows and 2 variables
}
\examples{
head(gsea)

}
\keyword{datasets}
