% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2m.R
\name{h2m}
\alias{h2m}
\title{how to make herb to molecule or cid of molecule}
\usage{
h2m(x, type = "latin", OB = 30, DL = 0.18)
}
\arguments{
\item{x}{the name of drug}

\item{type}{the type of drug name}

\item{OB}{the oral bioavailability of the compound}

\item{DL}{the drug-like properties of compounds}
}
\value{
molecule and cid
}
\description{
how to make herb to molecule or cid of molecule
}
\examples{
.h2m(x='Ziziphi Spinosae Semen',type='latin')
.h2m(x='houpu',type='pinyin',OB=40,DL=0.2)
h2m(x=c('Ziziphi Spinosae Semen','Abri Herba'),type='latin')
}
