% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd2t.R
\name{hd2t}
\alias{hd2t}
\title{How to make herb and disease to target}
\usage{
hd2t(
  hb,
  Disease,
  htype = "latin",
  dtype = "diseaseID",
  tSource = "tcmsp",
  OB = 30,
  DL = 0.18
)
}
\arguments{
\item{hb}{the name of the drug}

\item{Disease}{the name of the disease}

\item{htype}{the type of the hb}

\item{dtype}{the type of the disease}

\item{tSource}{the source of drug targets}

\item{OB}{the oral bioavailability of the compound}

\item{DL}{the drug-like properties of compounds}
}
\value{
targets and other information
}
\description{
How to make herb and disease to target
}
\examples{
.hd2t(hb='mahuang',Disease='C0009375',htype='pinyin',dtype='diseaseID',tSource='tcmsp',OB=35,DL=0.20)
.hd2t(hb='mahuang',Disease='C0009375',htype='pinyin',dtype='diseaseID',tSource='pubchem')
.hd2t(hb='Abutili Semen',Disease='Colonic Neoplasms',htype='latin',dtype='disease')
hd2t(hb=c('mahuang','huangqi'),Disease='C0019209',htype='pinyin',dtype='diseaseID',tSource='pubchem',OB=40,DL=0.25)
hd2t(hb=c('mahuang','huangqi'),Disease='C0019209',htype='pinyin',dtype='diseaseID',tSource='tcmsp')
}
