% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg}
\alias{kegg}
\title{kegg}
\format{
An object of class \code{data.frame} with 43304 rows and 5 columns.
}
\source{
KEGG database
}
\usage{
kegg
}
\description{
The results of KEGG pathways and genes
A data frame with 43304 rows and 5 variables
}
\examples{
head(kegg)

}
\keyword{datasets}
