% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_analyzedDEG.R
\name{load_analyzedDEG}
\alias{load_analyzedDEG}
\title{Load Analyzed DEG Results}
\usage{
load_analyzedDEG(id)
}
\arguments{
\item{id}{A vector for subset of `id` column in [AnalyzedDEG] or just a subset of [AnalyzedDEG].}
}
\value{
A `data.frame` or a list of `data.frame`.
}
\description{
Currently, the DEG results are for GSE85871.
}
\examples{

# Firstly, check the dataset
data("AnalyzedDEG")
head(AnalyzedDEG)

# Pick up what you want to load
# A subset of AnalyzedDEG is recommended,
# you can use \%>\%
# Otherwise specify a subset of id column

# e.g. head 5 sets
head5_reports <- AnalyzedDEG \%>\%
  head(n = 5) \%>\%
  load_analyzedDEG()

# only the second
one_report <- load_analyzedDEG(2)
}
