% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_analyzedsigpath.R
\name{load_analyzedsigpath}
\alias{load_analyzedsigpath}
\title{Load Analyzed Pathways Results}
\usage{
load_analyzedsigpath(id)
}
\arguments{
\item{id}{A vector for subset of \code{id} column in \link{Analyzedsigpath} or just a subset of \link{Analyzedsigpath}.}
}
\value{
A \code{data.frame} or a list of \code{data.frame}.
}
\description{
Currently, the top ten pathways results are for GSE85871.
}
\examples{

# Firstly, check the dataset
data("Analyzedsigpath")
head(Analyzedsigpath)

# Pick up what you want to load
# A subset of Analyzedsigpath is recommended,
# you can use \%>\%
# Otherwise specify a subset of id column

# e.g. head 5 sets
head5_reports <- Analyzedsigpath \%>\%
  head(n = 5) \%>\%
  load_analyzedsigpath()

# only the second
one_report <- load_analyzedDEG(2)

}
