% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_analyzedSigPathway.R
\name{load_AnalyzedSigPathway}
\alias{load_AnalyzedSigPathway}
\title{Load Analyzed Pathways Results}
\usage{
load_AnalyzedSigPathway(id)
}
\arguments{
\item{id}{A vector for subset of \code{id} column in \link{AnalyzedSigPathway} or just a subset of \link{AnalyzedSigPathway}.}
}
\value{
A \code{data.frame} or a list of \code{data.frame}.
}
\description{
Currently, the top ten pathways results are for GSE85871.
}
\examples{

# Firstly, check the dataset
data("AnalyzedSigPathway")
head(AnalyzedSigPathway)

# Pick up what you want to load
# A subset of AnalyzedSigPathway is recommended,
# you can use \%>\%
# Otherwise specify a subset of id column

# e.g. head 5 sets
head5_reports <- AnalyzedSigPathway \%>\%
  head(n = 5) \%>\%
  load_AnalyzedSigPathway()

# only the second
one_report <- load_AnalyzedSigPathway(2)
}
