% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_dataset.R
\name{load_example_dataset}
\alias{load_example_dataset}
\title{Load Example Dataset for Analysis}
\usage{
load_example_dataset(id = "GSE85871")
}
\arguments{
\item{id}{A dataset identifier.}
}
\value{
a \code{list}.
}
\description{
Load Example Dataset for Analysis
}
\examples{
x <- load_example_dataset()
head(x$expr)
head(x$pdata)

}
