% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_deg.R
\name{load_example_deg}
\alias{load_example_deg}
\title{Load all deg}
\usage{
load_example_deg(id = "GSE85871")
}
\arguments{
\item{id}{A dataset identifier}
}
\value{
a \code{list}.
}
\description{
Load all deg
}
\examples{
deg.all <- load_example_deg()

}
