% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2t.R
\name{m2t}
\alias{m2t}
\title{how to make molecule to symbol and fullname}
\usage{
m2t(x, type = "cid", tSource = "tcmsp")
}
\arguments{
\item{x}{is the name or cid of molecule}

\item{type}{is the type of x}

\item{tSource}{is the source of target}
}
\value{
symbol and fullname
}
\description{
how to make molecule to symbol and fullname
}
\examples{
.m2t(x='FER',type='molecule',tSource="tcmsp")
.m2t(x='FER',type='molecule',tSource="pubchem")
.m2t(x='445858',type='cid')
.m2t(x='MOL002547',type='molecule_id',tSource="tcmsp")
.m2t(x='MOL002547',type='molecule_id',tSource="pubchem")
 m2t(x=c('FER','pyrene'),type='molecule')
 m2t(x=c('445858','31423'),type='cid',tSource="tcmsp")
 m2t(x=c('445858','31423'),type='cid',tSource="pubchem")
 m2t(x=c('MOL002547','MOL000360'),type='molecule_id',tSource="tcmsp")
 m2t(x=c('MOL002547','MOL000360'),type='molecule_id',tSource="pubchem")

}
