% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_tme}
\alias{signature_tme}
\title{Information of signature_tme}
\format{
A list
}
\source{
See \verb{data-raw/}
}
\description{
Information of signature_tme
}
\examples{
data("signature_tme")
}
