% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_cor.R
\name{ss_cor}
\alias{ss_cor}
\title{Title Cor method}
\usage{
ss_cor(input, data, method)
}
\arguments{
\item{input}{a data.frame contains logFC value, which rownames should be geme symbols}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method from 'pearson',"spearman","kendall"}
}
\value{
a data.frame
}
\description{
Title Cor method
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
aa <- ss_cor(query2, data_logFC, method = "pearson")
aaa <- ss_cor(query2, data_logFC, method = "spearman")
aaaa <- ss_cor(query2, data_logFC, method = "kendall")
}
