% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_CoreGx.R
\name{ss_CoreGx}
\alias{ss_CoreGx}
\title{CoreGx method for Signature Search}
\usage{
ss_CoreGx(input, data, cores = 1L)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A \code{data.frame}.
}
\description{
CoreGx method for Signature Search
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
txp <- ss_CoreGx(query2, data_logFC[1:3])
}
