% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_lincs.R
\name{ss_lincsscore}
\alias{ss_lincsscore}
\title{Title Lincs methods}
\usage{
ss_lincsscore(input, data, method)
}
\arguments{
\item{input}{input is a list contains up and down genes}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method is lincs}
}
\value{
a data.frame
}
\description{
Title Lincs methods
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
lincs_kk <- ss_lincsscore(input, data_logFC, method = "lincs")
}
