% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string}
\alias{string}
\title{string}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\source{
\url{https://string-db.org/}
}
\usage{
string
}
\description{
Protein-protein interaction analysis was performed using STRING database.
A data frame with 2234 rows and 3 variables
}
\examples{
string.data <- data(string)
head(string.data)

}
\keyword{datasets}
