% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tcm.CircleHeatmap}
\alias{tcm.CircleHeatmap}
\title{Gene Expression Plot}
\usage{
tcm.CircleHeatmap(
  data,
  num = 3,
  colors = c("blue", "white", "red"),
  shownames = TRUE
)
}
\arguments{
\item{data}{Expression data of compounds}

\item{num}{Numbers of cluster}

\item{colors}{Color to show expression level}

\item{shownames}{Logic value to determine whether to display the gene name}
}
\value{
A plot
}
\description{
Gene Expression Plot
}
\examples{
tcm.CircleHeatmap(data = as.matrix(data_logFC[1:6, ]))
tcm.CircleHeatmap(data = as.matrix(data_logFC[1:6, ]), colors = c("blue", "white", "red"))
tcm.CircleHeatmap(data = as.matrix(data_logFC[1:10, ]), colors = c("skyblue", "white", "red"))
}
