% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tcm.EnrichBarplot}
\alias{tcm.EnrichBarplot}
\title{Barplot for enrichResult}
\usage{
tcm.EnrichBarplot(
  data,
  color = c("blue", "red"),
  linetype = "dashed",
  linecol = "chocolate",
  pointcol = "black",
  filcol = "red",
  colby = "pvalue"
)
}
\arguments{
\item{data}{enrichResult result in \code{data.frame} format}

\item{color}{Color to show}

\item{linetype}{Linetype}

\item{linecol}{Line color}

\item{pointcol}{Point color}

\item{filcol}{Bar color}

\item{colby}{Mapping colors by pvalue or p.adjust}
}
\value{
A \code{ggplot} object
}
\description{
Barplot for enrichResult
}
\examples{
data("AnalyzedSigPathway")
one_report <- tcm.LoadAnalyzedSigPathway(1)
tcm.EnrichBarplot(one_report)
}
