% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tcm.EnrichDotplot}
\alias{tcm.EnrichDotplot}
\title{Dotplot of enrichResult}
\usage{
tcm.EnrichDotplot(data, fill = "pvalue", color = c("blue", "red"))
}
\arguments{
\item{data}{enrichResult object which is \code{data.frame} format.}

\item{fill}{dot color}

\item{color}{color to show}
}
\value{
A \code{ggplot} object
}
\description{
Dotplot of enrichResult
}
\examples{
data("AnalyzedSigPathway")
one_report <- tcm.LoadAnalyzedSigPathway(1)
tcm.EnrichDotplot(one_report)
}
