% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{tcm.EnrichGO}
\alias{tcm.EnrichGO}
\title{Get Enriched GO Sets}
\usage{
tcm.EnrichGO(
  data,
  p.cutoff = 0.05,
  p.adj.cutoff = NULL,
  sortby = "pvalue",
  ont = "BP",
  n.path = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} from a subset of \link{tcm.RunDEG} result.}

\item{p.cutoff}{the p value cutoff.}

\item{p.adj.cutoff}{the cutoff value of p.adjust.}

\item{sortby}{A string value to indicate the pathway was sorted by p value or p.adjust value. Allowed value contains c('pvalue', 'p.adjust').}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{n.path}{A integer value to indicate how many top pathways sorted by p value should be identified; 10 by default.}
}
\value{
A \code{data.frame}.
}
\description{
Get Enriched GO Sets
}
\examples{
data("AnalyzedDEG")
one_report <- tcm.LoadAnalyzedDEG(1)
deg <- subset(one_report, abs(logFC) > 1 & P.Value < 0.05)
go <- tcm.EnrichGO(deg, p.cutoff = 0.05, p.adj.cutoff = 0.5, n.path = 10)
}
