% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{tcm.EnrichPathway}
\alias{tcm.EnrichPathway}
\title{Get Enriched Signal Pathways}
\usage{
tcm.EnrichPathway(
  data,
  p.cutoff = 0.05,
  p.adj.cutoff = 0.2,
  sortby = "pvalue",
  n.path = 10
)
}
\arguments{
\item{data}{A \code{data.frame} from a subset of \link{tcm.RunDEG} result.}

\item{p.cutoff}{the p value cutoff.}

\item{p.adj.cutoff}{the cutoff value of p.adjust.}

\item{sortby}{A string value to indicate the pathway was sorted by p value or p.adjust value. Allowed value contains c('pvalue', 'p.adjust').}

\item{n.path}{A integer value to indicate how many top pathways sorted by p value should be identified; 10 by default.}
}
\value{
A \code{data.frame}.
}
\description{
Get Enriched Signal Pathways
}
\examples{
data("AnalyzedDEG")
one_report <- tcm.LoadAnalyzedDEG(1)
deg <- subset(one_report, abs(logFC) > 1 & P.Value < 0.05)
path <- tcm.EnrichPathway(deg, p.cutoff = 1, p.adj.cutoff = 1, n.path = 10)
}
