% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-likeness.R
\name{tcm.GetDrugLikeness}
\alias{tcm.GetDrugLikeness}
\title{Druglikeness analysis}
\usage{
tcm.GetDrugLikeness(smiles)
}
\arguments{
\item{smiles}{SMILES of compounds}
}
\value{
a data frame
}
\description{
Druglikeness analysis
}
\examples{
data("SMILES")
tcm.GetDrugLikeness(SMILES)
}
