% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-similarity.R
\name{tcm.GetDrugSimilarity}
\alias{tcm.GetDrugSimilarity}
\title{Similarity calculation between two drugs}
\usage{
tcm.GetDrugSimilarity(smile_a, smile_b, method = "fp", plot = T)
}
\arguments{
\item{smile_a}{the smile of drug a}

\item{smile_b}{the smile of drug b}

\item{method}{the method for calculation}

\item{plot}{a logistical value}
}
\value{
a number
}
\description{
Similarity calculation between two drugs
}
\examples{
data("SMILES")
tcm.GetDrugSimilarity(smile_a = SMILES[1, ], smile_b = SMILES[2, ], method = "fp", plot = TRUE)
}
