% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-comparison.R
\name{tcm.GetDrugSpecificPathway}
\alias{tcm.GetDrugSpecificPathway}
\title{Drug specific pathway analysis}
\usage{
tcm.GetDrugSpecificPathway(
  index = c(1, 3, 5),
  data = data_logFC,
  group = "drug",
  num = 10,
  colorby = "pvalue",
  res.path = getwd(),
  plot = T
)
}
\arguments{
\item{index}{the index of colunm in data_logFC matrix}

\item{group}{the compare variable including group and class}

\item{num}{the number of pathways to show}

\item{colorby}{Mapping colors by pvalue or p.adjust}

\item{res.path}{A string value to indicate the path for saving the results for functional pathways.}

\item{plot}{A logic value to indicate if to plot pathways; TRUE by default.}
}
\value{
data frame
}
\description{
Drug specific pathway analysis
}
\examples{
data(data_logFC)
tcm.GetDrugSpecificPathway(
  group = "drug",
  data = data_logFC
)
}
