% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{tcm.LoadAnalyzedDEG}
\alias{tcm.LoadAnalyzedDEG}
\title{Load Analyzed DEG Results}
\usage{
tcm.LoadAnalyzedDEG(id)
}
\arguments{
\item{id}{A vector for subset of \code{id} column in \link{AnalyzedDEG} or just a subset of \link{AnalyzedDEG}.}
}
\value{
A \code{data.frame} or a list of \code{data.frame}.
}
\description{
Currently, the DEG results are for GSE85871.
}
\examples{

# Firstly, check the dataset
data("AnalyzedDEG")
head(AnalyzedDEG)

# Pick up what you want to load
# A subset of AnalyzedDEG is recommended,
# you can use \%>\%
# Otherwise specify a subset of id column

# e.g. head 5 sets
head5_reports <- AnalyzedDEG \%>\%
  head(n = 5) \%>\%
  tcm.LoadAnalyzedDEG()

# only the second
one_report <- tcm.LoadAnalyzedDEG(2)
}
