% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{tcm.LoadData}
\alias{tcm.LoadData}
\title{Load builtin/remote dataset}
\usage{
tcm.LoadData(name, just_query = FALSE)
}
\arguments{
\item{name}{a dataset name, e.g., "brca_disease", "TCM_expr", "tfdata".
Check \url{https://zenodo.org/record/5336709} for remote data.}

\item{just_query}{if \code{TRUE}, (download and) load data. Only set \code{TRUE} for
data in \code{rda} format.}
}
\value{
a data or a string represting data path.
}
\description{
Load builtin/remote dataset
}
\examples{
tcm.LoadData("SMILES")
}
