% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{tcm.LoadExampleDataset}
\alias{tcm.LoadExampleDataset}
\title{Load Example Dataset for Analysis}
\usage{
tcm.LoadExampleDataset(id = "GSE85871")
}
\arguments{
\item{id}{A dataset identifier.}
}
\value{
a \code{list}.
}
\description{
Load Example Dataset for Analysis
}
\examples{
x <- tcm.LoadExampleDataset()
head(x$expr)
head(x$pdata)
}
