% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tcm.RunDEG}
\alias{tcm.RunDEG}
\alias{tcm.RunBatchDEG}
\title{Differential Expression Analysis}
\usage{
tcm.RunDEG(data, group, ref_group)

tcm.RunBatchDEG(data, groups, ref_group)
}
\arguments{
\item{data}{An expression dataset in \code{data.frame} format whose rows indicate genes and columns indicate samples.}

\item{group}{A character vector specifying 1 of 2 groups which samples belong to.}

\item{ref_group}{A string specifying the reference group.}

\item{groups}{A character vector specifying the group which samples belong to. Should at least 2 groups.}
}
\value{
A \code{data.frame} or a \code{list} of \code{data.frame}.
}
\description{
The functions are designed to run DEA.
\itemize{
\item \code{tcm.RunDEG} is used to run two group differential expression analysis.
\item \code{tcm.RunBatchDEG} is used to run batch DEG analysis.
}
}
\examples{
data <- tcm.LoadExampleDataset()

# Get a subset of expression dataset
ix <- c(1, 2, 61, 62)
expr <- data$expr[, ix]
group <- data$pdata$perturbagen[ix]

# Run DEG analysis
report <- tcm.RunDEG(expr, group = group, ref_group = group[3])
head(report)

# Run batch DEG analysis
reports <- tcm.RunBatchDEG(expr, groups = group, ref_group = group[3])
head(reports)

identical(report, reports[[1]])
}
