% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{tcm.SeekDrugPair}
\alias{tcm.SeekDrugPair}
\title{Drug Pair Seeker}
\usage{
tcm.SeekDrugPair(data = NULL, disease = NULL, cutoff = 1)
}
\arguments{
\item{data}{the logFC matrix of TCM}

\item{disease}{the disease of you intersted}

\item{cutoff}{DEG cutoff}
}
\value{
the data.frame
}
\description{
Drug Pair Seeker
}
\examples{
data("data_logFC")
data("tcga_disease")
tmp <- tcm.SeekDrugPair(data = data_logFC, disease = res_RNA2)
}
