% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithCMAP}
\alias{tcm.SSwithCMAP}
\title{CMAP method for Signature Search}
\usage{
tcm.SSwithCMAP(input, data, cores = 1L)
}
\arguments{
\item{input}{A list containing UP (in \code{upset} element) and DOWN (in \code{downset} element) genes.}

\item{data}{A \code{data.frame} containing \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A data.frame.
}
\description{
CMAP method for Signature Search
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
cmap_kk <- tcm.SSwithCMAP(input = input, data = data_logFC)
}
