% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithCoreGx}
\alias{tcm.SSwithCoreGx}
\title{CoreGx method for Signature Search}
\usage{
tcm.SSwithCoreGx(input, data, cores = 1L)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A \code{data.frame}.
}
\description{
CoreGx method for Signature Search
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
txp <- tcm.SSwithCoreGx(query2, data_logFC[1:10])
}
