% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithCorrelation}
\alias{tcm.SSwithCorrelation}
\title{Cor method for for Signature Search}
\usage{
tcm.SSwithCorrelation(
  input,
  data,
  method = c("pearson", "spearman", "kendall")
)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{method}{A string specifying correlation calculation method, should be one of "pearson", "spearman", "kendall".}
}
\value{
A \code{data.frame}.
}
\description{
Cor method for for Signature Search
}
\examples{
data("data_logFC")
query <- data_logFC[1:60, 1, drop = FALSE]
aa <- tcm.SSwithCorrelation(query, data_logFC, method = "pearson")
aaa <- tcm.SSwithCorrelation(query, data_logFC, method = "spearman")
aaaa <- tcm.SSwithCorrelation(query, data_logFC, method = "kendall")
}
