% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithIntegration}
\alias{tcm.SSwithIntegration}
\title{tcm.SSwithIntegration method for Signature Search}
\usage{
tcm.SSwithIntegration(
  input,
  data,
  methodslist = list("CMAP", "Correlation", "GCMAP", "LINCS", "ZS", "XSum", "XCos"),
  type = "spearman",
  threshold = 1,
  direction = "down",
  top = 10,
  cores = 1L
)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A data.frame.
}
\description{
tcm.SSwithIntegration method for Signature Search
}
\examples{
data <- data_logFC
input <- data_logFC[, 1, drop = FALSE]
tcm.SSwithIntegration(input,data)
}
