% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithLINCS}
\alias{tcm.SSwithLINCS}
\title{Lincs method for Signature Search}
\usage{
tcm.SSwithLINCS(input, data)
}
\arguments{
\item{input}{A list containing UP (in \code{upset} element) and DOWN (in \code{downset} element) genes.}

\item{data}{A \code{data.frame} containing \code{logFC} value data from 103 compounds.}
}
\value{
A data.frame.
}
\description{
Lincs method for Signature Search
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
lincs_kk <- tcm.SSwithLINCS(input, data_logFC)
}
