% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithXCos}
\alias{tcm.SSwithXCos}
\title{tcm.SSwithXCos method for Signature Search}
\usage{
tcm.SSwithXCos(input, data, cores = 1L)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A data.frame.
}
\description{
tcm.SSwithXCos method for Signature Search
}
