% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-search.R
\name{tcm.SSwithXSum}
\alias{tcm.SSwithXSum}
\title{tcm.SSwithXSum method for Signature Search}
\usage{
tcm.SSwithXSum(input, data, cores = 1L)
}
\arguments{
\item{input}{A \code{data.frame} containing \code{logFC} value, which \code{rownames} should be gene symbols.}

\item{data}{A \code{data.frame} contains \code{logFC} value data from 103 compounds.}

\item{cores}{Number of cores to run the task}
}
\value{
A data.frame.
}
\description{
tcm.SSwithXSum method for Signature Search
}
\examples{
data <- data_logFC
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
tcm.SSwithXSum(input,data)
}
