% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_prescription.R
\name{tcm_prescription}
\alias{tcm_prescription}
\title{Find Chinese medicine prescriptions through disease targets}
\usage{
tcm_prescription(
  disease.gene,
  compound.source = "tcmp",
  OB.value = 30,
  DL.value = 0.18,
  freq.value = 500,
  digits.value = 2,
  ...
)
}
\arguments{
\item{disease.gene}{disease gene}

\item{compound.source}{"tcmp", "cmp"}

\item{OB.value}{bioavailability}

\item{DL.value}{drug-likeness}

\item{freq.value}{Minimum number of occurrences of herbs}

\item{digits.value}{number of digits after decimal point}

\item{...}{additional parameters}
}
\value{
list:
\enumerate{
\item Frequency of occurrence of herbs
\item Selected traditional Chinese medicine prescription results
}
}
\description{
Find Chinese medicine prescriptions through disease targets
}
\examples{
\dontrun{
data(disease_gene, package = "TCMR")
newdata <- tcm_prescription(disease_gene)
newdata[[1]]
newdata[[2]]
}
}
