% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCMR-package.R
\docType{package}
\name{TCMR-package}
\alias{TCMR}
\alias{TCMR-package}
\title{TCMR: Traditional Chinese Medicine Network Pharmacology}
\description{
A visualization tool for network pharmacology data analysis.
}
\author{
\strong{Maintainer}: Jinkun Liu \email{liujinku007@163.com} (\href{https://orcid.org/0000-0003-2212-9617}{ORCID})

}
\keyword{internal}
