% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichtf.R
\name{tf_enrich}
\alias{tf_enrich}
\title{Title Tf enrichment for genes}
\usage{
tf_enrich(
  gene,
  tfdata = tfdata,
  colby = "pvalue",
  res.path = getwd(),
  colours = c("red", "yellow"),
  n = 20
)
}
\arguments{
\item{gene}{The special gene for enrichment}

\item{tfdata}{the transcription factor and gene correspondence table}

\item{colby}{mapping colors by pvalue or p.adjust}

\item{res.path}{A string value to indicate the path for saving the results for functional pathways.}

\item{colours}{color to show}

\item{n}{A integer value to indicate how many top tfs sorted by GeneRatio should be identified; 20 by default.}
}
\value{

}
\description{
Title Tf enrichment for genes
}
\examples{
data("tfdata")
gene <- tfdata$gene[100:500]
tf_enrich(gene,tfdata = tfdata,colours=c("red","blue"),colby = "p.adjust")
tf_enrich(gene,tfdata = tfdata,colours=c("red","blue"),colby = "pvalue")
}
