% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateDrugID.R
\name{translateDrugID}
\alias{translateDrugID}
\title{translateDrugID}
\usage{
translateDrugID(query, from, to)
}
\arguments{
\item{query}{the identifer of compounds in other database}

\item{from}{the first}

\item{to}{the second}
}
\value{
a data.frame
}
\description{
translateDrugID
}
\details{
Interface to CTS (http://cts.fiehnlab.ucdavis.edu/) for metabolite identifier translation between200 of the most common biological databases including: Chemical Name, InChIKey, PubChem CID,ChemSpider, BioCyc, ChEBI, CAS, HMDB, KEGG and LipidMAPS.
}
\examples{
query<-c("C15973","C00026")
from<-"KEGG"
to<-"PubChem CID"
translateDrugID(query,from,to)
}
