% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trrust}
\alias{trrust}
\title{TRRUST database}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16453 rows and 6 columns.
}
\source{
\url{https://www.grnpedia.org/trrust/}
}
\usage{
trrust
}
\description{
TRRUST contains 8,444 and 6,552 TF-target regulatory
relationships of 800 human TFs and 828 mouse TFs, respectively.

The datasets was from TRRUST database
A data frame with 16543 rows and 6 variables
}
\examples{
head(trrust)
}
\keyword{datasets}
