% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venn.data}
\alias{venn.data}
\title{Venn diagram example data}
\format{
An object of class \code{data.frame} with 6439 rows and 2 columns.
}
\source{
\url{https://www.nature.com/articles/s41556-021-00796-6#Sec47}
}
\usage{
venn.data
}
\description{
The datasets was from xfbdf and https://doi.org/10.1038/s41556-021-00796-6.
A data frame with 6439 rows and 2 variables
}
\examples{
head(venn.data)
}
\keyword{datasets}
