% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Barplot for enrichResult}
\usage{
bar_plot(
  data,
  color = c("blue", "red"),
  linetype = "dashed",
  linecol = "chocolate",
  pointcol = "black",
  filcol = "red",
  colby = "pvalue"
)
}
\arguments{
\item{data}{enrichResult result in \code{data.frame} format}

\item{color}{Color to show}

\item{linetype}{Linetype}

\item{linecol}{Line color}

\item{pointcol}{Point color}

\item{filcol}{Bar color}

\item{colby}{Mapping colors by pvalue or p.adjust}
}
\value{
A \code{ggplot} object
}
\description{
Barplot for enrichResult
}
\examples{
data("AnalyzedSigPathway")
one_report <- load_AnalyzedSigPathway(1)
bar_plot(one_report)
}
