% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Barplot of enrichResult}
\usage{
bar_plot(
  data,
  color = c("blue", "red"),
  linetype = "dashed",
  linecol = "chocolate",
  pointcol = "black",
  filcol = "red",
  colby = "pvalue"
)
}
\arguments{
\item{data}{enrichResult object which is `data.frame` format.}

\item{color}{color to show}

\item{linetype}{linetype}

\item{linecol}{line color}

\item{pointcol}{point color}

\item{filcol}{bar color}

\item{colby}{mapping colors by pvalue or p.adjust}
}
\value{
A `ggplot` object
}
\description{
Barplot of enrichResult
}
\examples{

data("AnalyzedSigPathway")
one_report <- load_AnalyzedSigPathway(1)
bar_plot(one_report)

}
