% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cmap}
\alias{cmap}
\title{Title CMAP methods}
\usage{
cmap(input, data, method)
}
\arguments{
\item{input}{input is a list contains up and down genes}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method is cmap}
}
\value{
a data.frame
}
\description{
Title CMAP methods
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
cmap_kk <- cmap(input = input, data = data_logFC, method = "CMAP")
}
