% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_dot.R
\name{dotplot}
\alias{dotplot}
\title{Dotplot of enrichResult}
\usage{
dotplot(data, fill = "pvalue", color = c("blue", "red"))
}
\arguments{
\item{data}{enrichResult object which is \code{data.frame} format.}

\item{fill}{dot color}

\item{color}{color to show}
}
\value{
A \code{ggplot} object
}
\description{
Dotplot of enrichResult
}
\examples{
data("AnalyzedSigPathway")
one_report <- load_AnalyzedSigPathway(1)
dotplot(one_report)
}
