% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dot.R
\name{gene_dot}
\alias{gene_dot}
\title{Dotplot of Special Genes in Special Compounds of logFC Matrix}
\usage{
gene_dot(data, color = c("red", "blue"))
}
\arguments{
\item{data}{Special genes expression data of special compounds}

\item{color}{Color to distinguish between different trends}
}
\value{
A `ggplot` object
}
\description{
Dotplot of Special Genes in Special Compounds of logFC Matrix
}
\examples{
data <- as.matrix(data_logFC[1:6, ])
data <- data.frame(t(data), check.rows = FALSE, check.names = FALSE)
data$compound <- rownames(data)
data <- data[3:13, ]
gene_dot(data)
}
