% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneplot.R
\name{geneplot}
\alias{geneplot}
\title{Gene Expression Plot}
\usage{
geneplot(data, num = 3, colors = c("blue", "white", "red"))
}
\arguments{
\item{data}{Expression data of compounds}

\item{num}{Numbers of cluster}

\item{colors}{Color to show expression level}
}
\value{
A plot
}
\description{
Gene Expression Plot
}
\examples{
geneplot(data = as.matrix(data_logFC[1:6, ]))
geneplot(data = as.matrix(data_logFC[1:6, ]), colors = c("blue", "white", "red"))
geneplot(data = as.matrix(data_logFC[1:6, ]), colors = c("skyblue", "white", "red"))
}
