% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneplot.R
\name{geneplot}
\alias{geneplot}
\title{Title gene expression plot}
\usage{
geneplot(data, num = 3, colors = c("skyblue", "white", "red"))
}
\arguments{
\item{data}{the expression data of compounds}

\item{num}{the numbers of cluster}

\item{colors}{the color to show expression level}
}
\value{

}
\description{
Title gene expression plot
}
