% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdotplot.R
\name{getdotplot}
\alias{getdotplot}
\title{Dotplot of enrichResult}
\usage{
getdotplot(data, fill = "pvalue", color = c("blue", "red"))
}
\arguments{
\item{data}{enrichResult object which is \code{data.frame} format.}

\item{fill}{dot color}

\item{color}{color to show}
}
\value{
ggplot object
}
\description{
Dotplot of enrichResult
}
\examples{
data("Analyzedsigpath")
one_report <- load_analyzedsigpath(1)
getdotplot(one_report)
}
