% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd2t.R
\name{hd2t}
\alias{hd2t}
\title{How to make herb and disease to target}
\usage{
hd2t(hb, Disease, htype = "latin", dtype = "diseaseID")
}
\arguments{
\item{hb}{hb is the name of the drug}

\item{Disease}{is the name of the disease}

\item{htype}{is the type of the hb}

\item{dtype}{is the type of the disease}
}
\value{
targets and other information
}
\description{
How to make herb and disease to target
}
\examples{
hd2t(hb='mahuang',Disease='C0009375',htype='pinyin',dtype='diseaseID')
hd2t(hb='Abutili Semen',Disease='Colonic Neoplasms',htype='latin',dtype='disease')
}
