% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincs.R
\name{lincsscore}
\alias{lincsscore}
\title{Title Lincs methods}
\usage{
lincsscore(input, data, method)
}
\arguments{
\item{input}{input is a list contains up and down genes}

\item{data}{a data.frame contains logFC value data from 103 compounds}

\item{method}{a method is lincs}
}
\value{
a data.frame
}
\description{
Title Lincs methods
}
