% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_cmap.R
\name{ss_cmap}
\alias{ss_cmap}
\title{CMAP method for Signature Search}
\usage{
ss_cmap(input, data)
}
\arguments{
\item{input}{A list containing UP (in \code{upset} element) and DOWN (in \code{downset} element) genes.}

\item{data}{A \code{data.frame} containing \code{logFC} value data from 103 compounds.}
}
\value{
A data.frame.
}
\description{
CMAP method for Signature Search
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
cmap_kk <- ss_cmap(input = input, data = data_logFC)
}
