% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_cor.R
\name{ss_cor}
\alias{ss_cor}
\title{Cor method for for Signature Search}
\usage{
ss_cor(input, data, method = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{input}{A `data.frame` containing `logFC` value, which `rownames` should be gene symbols.}

\item{data}{A `data.frame` contains `logFC` value data from 103 compounds.}

\item{method}{A string specifying correlation calculation method, should be one of "pearson", "spearman", "kendall".}
}
\value{
A data.frame.
}
\description{
Cor method for for Signature Search
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
aa <- ss_cor(query2, data_logFC, method = "pearson")
aaa <- ss_cor(query2, data_logFC, method = "spearman")
aaaa <- ss_cor(query2, data_logFC, method = "kendall")
}
