% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_CoreGx.R
\name{ss_CoreGx}
\alias{ss_CoreGx}
\title{Title CoreGx method}
\usage{
ss_CoreGx(input, data)
}
\arguments{
\item{input}{a data.frame contains logFC value, which rownames should be geme symbols}

\item{data}{a data.frame contains logFC value data from 103 compounds}
}
\value{
a data.frame
}
\description{
Title CoreGx method
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
txp <- ss_CoreGx(query2, data_logFC[1:3])
}
