% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_gcmap.R
\name{ss_gcmap}
\alias{ss_gcmap}
\title{gcmap Method for Signature Search}
\usage{
ss_gcmap(input, data, higher = 1, lower = -1)
}
\arguments{
\item{input}{A `data.frame` containing `logFC` value, which `rownames` should be gene symbols.}

\item{data}{A `data.frame` contains `logFC` value data from 103 compounds.}

\item{higher}{A cutoff value. If `logFC` larger than or equal to `higher` will be included
in the gene set with  `+1`.}

\item{lower}{A cutoff value. If `logFC` smaller than or equal to `higher` will be included
in the gene set to `-1`, others should be set to `0`.}
}
\value{
A `data.frame`.
}
\description{
gcmap Method for Signature Search
}
\examples{
data("data_logFC")
query2 <- data_logFC[1:60, 1, drop = FALSE]
gcmap_kk <- ss_gcmap(input = query2, data = data_logFC)
}
