% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_lincs.R
\name{ss_lincsscore}
\alias{ss_lincsscore}
\title{Lincs method for Signature Search}
\usage{
ss_lincsscore(input, data)
}
\arguments{
\item{input}{A list containing UP (in `upset` element) and DOWN (in `downset` element) genes.}

\item{data}{A `data.frame` containing `logFC` value data from 103 compounds.}
}
\value{
A data.frame.
}
\description{
Lincs method for Signature Search
}
\examples{
data("data_logFC")
upset <- rownames(data_logFC)[1:100]
downset <- rownames(data_logFC)[400:550]
input <- list(upset = upset, downset = downset)
lincs_kk <- ss_lincsscore(input, data_logFC)
}
