% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCMR-package.R
\docType{package}
\name{TCMR-package}
\alias{TCMR}
\alias{TCMR-package}
\title{TCMR: Traditional Chinese Medicine Research Toolbox}
\description{
Provides functions for analyzing traditional Chinese drugs in gene expression arrays.
}
\author{
\strong{Maintainer}: Saisai Tian \email{saisaitian@somewhere.net}

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID}) [contributor]
}

}
\keyword{internal}
