% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_enrich.R
\name{tf_enrich}
\alias{tf_enrich}
\title{TF Enrichment for Genes}
\usage{
tf_enrich(
  gene,
  tfdata = tfdata,
  colby = "pvalue",
  res.path = getwd(),
  colours = c("red", "yellow"),
  n = 20
)
}
\arguments{
\item{gene}{The special gene for enrichment}

\item{tfdata}{Transcription factor and gene correspondence table}

\item{colby}{Mapping colors by \code{pvalue} or \code{p.adjust}}

\item{res.path}{A string value to indicate the path for saving the results for functional pathways.}

\item{colours}{Color to show}

\item{n}{An integer value to indicate how many top tfs sorted by GeneRatio should be identified; 20 by default.}
}
\value{
A \code{ggplot} object
}
\description{
TF Enrichment for Genes
}
\examples{
data("tfdata")
gene <- tfdata$gene[100:500]
tf_enrich(gene, tfdata = tfdata, colours = c("red", "blue"), colby = "p.adjust")
tf_enrich(gene, tfdata = tfdata, colours = c("red", "blue"), colby = "pvalue")
}
